requirements.txt
selenium
pytest
pytest-html
webdriver-manager
pyyaml
